--- slack.ts
@@ -1,16 +1,0 @@
-import { Run, ExecuteOptions, executeInternal } from "./tasks";
-
-export default {
-  message: async <Output = undefined | null>(
-    channelName: string,
-    message: string,
-    opts?: ExecuteOptions
-  ): Promise<Run<Record<string, unknown> | undefined | null, Output>> => {
-    return executeInternal(
-      "airplane:slack_message",
-      { channelName, message },
-      { slack: "res00000000zteamslack" },
-      opts
-    );
-  },
-};
