+++ src/index.ts
@@ -1,0 +1,4 @@
+export { appendOutput, setOutput } from "./internal/output";
+export { execute } from "./internal/execute";
+export * from "./internal/builtins";
+export { ClientOptions } from "./internal/api/client";
