+++ src/internal/builtins/index.ts
@@ -1,0 +1,5 @@
+export * as email from "./email";
+export * as mongodb from "./mongodb";
+export * as rest from "./rest";
+export * as slack from "./slack";
+export * as sql from "./sql";
