+++ src/internal/output.test.ts
