+++ src/internal/output.ts
